+++ lib/errors/UnknownEventNameError.js
@@ -1,0 +1,9 @@
+class UnknownEventNameError extends Error {
+  constructor(record) {
+    super(`"${record.eventName}" is an unknown event name`);
+    this.name = 'UnknownEventNameError';
+    this.details = record;
+  }
+}
+
+module.exports = UnknownEventNameError;
