+++ lib/errors/index.js
@@ -1,0 +1,9 @@
+const FieldNotFoundError = require('./FieldNotFoundError');
+const UnknownEventNameError = require('./UnknownEventNameError');
+const ValidationError = require('./ValidationError');
+
+module.exports = {
+  FieldNotFoundError,
+  UnknownEventNameError,
+  ValidationError
+};
